+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* **Breaking change** The `EnableConnectionLevelAttributes` option has been removed.
+  ([#2414](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2414))
+
 ## 1.10.0-beta.1
 
 Released 2024-Dec-09