+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -15,6 +15,9 @@
   by a `?` character.
   ([#2446](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2446))
 
+* Updated OpenTelemetry core component version(s) to `1.11.0`.
+  ([#2470](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2470))
+
 ## 1.10.0-beta.1
 
 Released 2024-Dec-09